Updating name of commands, attempting to get user-profile to pull the name into the response.
