+++ events/interactionCreate.js
@@ -5,6 +5,10 @@ module.exports = {
     async execute(interaction) {
         if (!interaction.isChatInputCommand()) return;
 
+        // if (interaction.isApplicationCommand()) {
+
+        // }
+
         const command = interaction.client.commands.get(interaction.commandName);
 
         if (!command) {
